style: Move styling code into another js files on RecommendResourceContainer
